refactor, avoid copying BigtableIO
