+++ version.sbt
@@ -1,1 +1,1 @@
-version in ThisBuild := "0.5.3-SNAPSHOT"
+version in ThisBuild := "0.5.3-courage-bt"
